% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest_rule.R
\name{contest_rule}
\alias{contest_rule}
\title{Contest a rule}
\usage{
contest_rule(id, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{id of rule to contest.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
A `list` information about the rule contested.
}
\description{
Contest a rule
}
\examples{
\dontrun{
contest_rule(1)
}
}
