% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_rule.R
\name{delete_rule}
\alias{delete_rule}
\title{Delete a rule}
\usage{
delete_rule(id, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{the id of the rule}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
list of information about deleted rule.
}
\description{
Delete a rule
}
\examples{
\dontrun{
delete_rule(1)
}
}
