% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_ruleset.R
\name{delete_ruleset}
\alias{delete_ruleset}
\title{Delete a ruleset}
\usage{
delete_ruleset(id, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{the id of the ruleset}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
list of information about deleted ruleset
}
\description{
Delete a ruleset
}
\examples{
\dontrun{
delete_ruleset(1)
}
}
