% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{gbifrules_get_}
\alias{gbifrules_get_}
\title{gbifrules_get_}
\usage{
gbifrules_get_(url, query)
}
\arguments{
\item{url}{helper}

\item{query}{helper}
}
\description{
gbifrules_get_
}
\keyword{internal}
