% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{gbifrules_post}
\alias{gbifrules_post}
\title{gbifrules_post}
\usage{
gbifrules_post(url, body, user = NULL, pwd = NULL)
}
\arguments{
\item{url}{helper}

\item{body}{helper}

\item{user}{optional user for authentication, defaults to GBIF_USER env variable}

\item{pwd}{optional password for authentication, defaults to GBIF_PWD env variable}
}
\description{
gbifrules_post
}
\keyword{internal}
