% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_project.R
\name{make_project}
\alias{make_project}
\title{Make Project}
\usage{
make_project(name = NULL, description = NULL, user = NULL, pwd = NULL)
}
\arguments{
\item{name}{the name of the project.}

\item{description}{describe the project.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
The projectId of the project.
}
\description{
Make Project
}
\examples{
\dontrun{
make_project("Example project","An example Project")
}
}
