% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ruleset.R
\name{make_ruleset}
\alias{make_ruleset}
\title{Make Ruleset}
\usage{
make_ruleset(
  projectId = NULL,
  name = NULL,
  description = NULL,
  user = NULL,
  pwd = NULL
)
}
\arguments{
\item{projectId}{the id of the project the ruleset should belong to.}

\item{name}{the name of the project.}

\item{description}{describe the project.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
The rulesetId and projectId of the ruleset as a list.
}
\description{
Make Ruleset
}
\examples{
\dontrun{
make_ruleset(1,"Name of project","Example description")
}
}
