% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_contest_rule.R
\name{rm_contest_rule}
\alias{rm_contest_rule}
\title{Remove contest of a rule}
\usage{
rm_contest_rule(id, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{id of rule to contest.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
A `list` information about the rule contested.
}
\description{
Remove contest of a rule
}
\examples{
\dontrun{
rm_contest_rule(1)
}
}
