% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_support_rule.R
\name{rm_support_rule}
\alias{rm_support_rule}
\title{Remove support for a rule}
\usage{
rm_support_rule(id = NULL, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{the id of the rule to updated.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
A `list` information about the rule supported.
}
\description{
Remove support for a rule
}
\examples{
\dontrun{
rm_support_rule(1)
}
}
