% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_rule.R
\name{support_rule}
\alias{support_rule}
\title{Support a rule}
\usage{
support_rule(id = NULL, user = NULL, pwd = NULL)
}
\arguments{
\item{id}{the id of the rule to updated.}

\item{user}{(character) Optional username for authentication. Defaults to GBIF_USER environment variable.}

\item{pwd}{(character) Optional password for authentication. Defaults to GBIF_PWD environment variable.}
}
\value{
A `list` information about the rule supported.
}
\description{
Support a rule
}
\examples{
\dontrun{
support_rule(1)
}
}
