% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_download.R
\name{clean_download}
\alias{clean_download}
\title{Clean a GBIF download of suspicious records using complex rules}
\usage{
clean_download(d, rm_suspicious = TRUE, handle_conflicts = "favor_suspicious")
}
\arguments{
\item{d}{a GBIF download in DWCA format}

\item{rm_suspicious}{removes records with suspicious annotations}

\item{handle_conflicts}{how to handle conflicting annotations}
}
\value{
A cleaned download `data.frame()`
}
\description{
Removes records that have been marked as suspicious by annotation users using complex rules.
Only filters based on SUSPICIOUS annotations, ignoring other annotation types.
}
\examples{
\dontrun{
library(rgbif)
# End to end example with GBIF download and cleaning

# Ambystoma mexicanum
occ_download(
pred(taxonKey,"2431950"),
pred_default()
format = "SIMPLE_CSV"
)

# Download and clean
d <- occ_download_get('0051416-241126133413365') \%>\%
 occ_download_import()

clean_download(d) 
}


}
