% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_rule_comment.R
\name{delete_rule_comment}
\alias{delete_rule_comment}
\title{Delete a rule comment}
\usage{
delete_rule_comment(ruleId, commentId)
}
\arguments{
\item{ruleId}{The ID of the rule that the comment belongs to}

\item{commentId}{The ID of the comment to delete}
}
\value{
Nothing (invisible NULL)
}
\description{
Delete a rule comment
}
\examples{
\dontrun{
delete_rule_comment(ruleId = 1, commentId = 5)
}
}
