% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_helpers.R
\name{gbifrules_post}
\alias{gbifrules_post}
\title{gbifrules_post}
\usage{
gbifrules_post(url, body)
}
\arguments{
\item{url}{helper}

\item{body}{helper}
}
\description{
gbifrules_post
}
\keyword{internal}
